# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OCMS & Store LaPet Omnichannel Master API
  description: >
    Hợp đồng giao tiếp API chuẩn hóa toàn diện (Master OpenAPI Specification) giữa
    Hệ thống Quản trị Vận hành Đa kênh OCMS và Website Bán Hàng / Cổng Đại lý (store-lapet),
    đồng thời tích hợp liên thông với ERP Odoo và Kênh Zalo OA.
    
    Tài liệu này bao quát toàn bộ các nhóm tính năng vận hành:
    1. **Orders (Đơn hàng & Tiếp nhận chuẩn):** Cổng tiếp nhận đơn hàng bất đồng bộ, tra cứu trạng thái, duyệt đơn Odoo, hủy đơn và thống kê.
    2. **Products & Inventory (Sản phẩm & Tồn kho):** Danh mục sản phẩm, bảng giá đại lý/bán lẻ, kiểm tra tồn kho Odoo thời gian thực.
    3. **Promotions & Pricing Engine (Khuyến mãi & Động cơ tính giá):** Động cơ tính chiết khấu bậc thang tự động cho giỏ hàng, tư vấn ưu đãi bằng AI.
    4. **Customers & Dealers (Khách hàng & Đại lý B2B):** Xác thực OTP/Mật khẩu, hồ sơ đại lý, hạn mức công nợ, doanh số và quản lý danh bạ CRM.
    5. **Omnichannel Chat (Hội thoại & Zalo):** Quản lý luồng trò chuyện, gửi tin nhắn trao đổi, tự động bóc tách đơn hàng AI từ chat.
    6. **Webhooks & Sync (Thanh toán & Đồng bộ):** Webhook thanh toán tự động VietQR và kích hoạt đồng bộ dữ liệu Odoo ERP.
  version: 2.0.0
  contact:
    name: OCMS Integration Architecture Team
    email: dev@lapet.vn
servers:
  - url: http://localhost:3000
    description: Môi trường phát triển cục bộ (Local Development)
  - url: https://api.ocms.lapet.vn
...
paths:
...
  /api/v1/orders/standard:
    post:
      tags:
        - Orders
      summary: Cổng tiếp nhận đơn hàng chuẩn hóa từ Web Bán Hàng (Hàng đợi bất đồng bộ)
      description: >
        Nhận gói đơn hàng hoàn chỉnh từ Web Bán Hàng (store-lapet), kiểm duyệt dữ liệu tự động
        qua bộ kiểm duyệt Schema, ghi nhận trạng thái ban đầu `pending` vào CSDL OCMS, và ném gói việc
        vào Trục Hàng Đợi (Redis Queue) để Worker tạo Sale Order trên Odoo và gửi tin Zalo ngầm.
      operationId: createStandardOrder
      requestBody:
        required: true
        description: Dữ liệu đơn hàng đóng gói từ giỏ hàng và thanh toán trên Web
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StandardOrderPayload'
            example:
              order_code: "WEB-20260909-8A1B"
              partner_id: 17871
              customer_name: "Nguyễn Văn An"
              customer_phone: "0901234567"
              customer_email: "nguyenvanan@example.com"
              company_name: "Công Ty TNHH Thú Cưng PetCare"
              vat_number: "0312345678"
              shipping_address: "123 Đường Nguyễn Trãi, Phường 2, Quận 5, TP. Hồ Chí Minh"
              items:
                - sku: "PAT-LAMB-100"
                  product_name: "Pate Thịt Cừu Cho Chó 100g"
                  odoo_product_id: 1042
                  quantity: 10
                  price: 25000
                - sku: "BONE-BEEF-M"
                  product_name: "Xương Gặm Vị Bò Size M"
                  odoo_product_id: 1058
                  quantity: 2
                  price: 45000
              subtotal_amount: 340000
              discount_amount: 20000
              shipping_fee: 30000
              total_amount: 350000
              voucher_code: "LAPETHE2026"
              payment_method: "cod"
              note: "Giao trong giờ hành chính, gọi trước khi giao 15 phút"
      responses:
        '201':
          description: Đơn hàng hợp lệ và đã được đưa vào hàng đợi xử lý thành công (< 0.1s)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderCreatedResponse'
        '400':
          description: Dữ liệu đơn hàng sai định dạng hoặc thiếu các trường bắt buộc
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Lỗi máy chủ nội bộ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
  /api/v1/orders/standard/{order_code}:
    get:
      tags:
        - Orders
      summary: Tra cứu tiến độ & trạng thái đơn hàng thời gian thực
      description: >
        Cho phép Web Bán Hàng hoặc khách hàng tra cứu trạng thái đơn hàng theo mã `order_code`
        hoặc mã theo dõi `tracking_id`. Trả về trạng thái xử lý trên OCMS và liên kết đơn hàng Odoo nếu đã tạo thành công.
      operationId: getStandardOrderStatus
      parameters:
        - name: order_code
          in: path
          required: true
          description: Mã đơn hàng (ví dụ WEB-20260909-8A1B) hoặc UUID tracking_id
          schema:
            type: string
          example: "WEB-20260909-8A1B"
      responses:
        '200':
          description: Thông tin tiến độ đơn hàng
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderStatusDetailResponse'
        '404':
          description: Không tìm thấy đơn hàng với mã tương ứng
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    StandardOrderPayload:
      type: object
      required:
        - order_code
        - customer_name
        - customer_phone
        - shipping_address
        - items
        - total_amount
      properties:
        order_code:
          type: string
          description: Mã định danh duy nhất từ Web Bán Hàng (Idempotency Key chống gửi trùng)
          minLength: 3
          maxLength: 80
          example: "WEB-20260909-8A1B"
        partner_id:
          type: integer
          nullable: true
          description: ID đối tác Odoo ERP (`res.partner.id`) nếu là đại lý đã đăng nhập
          example: 17871
        customer_name:
          type: string
          description: Họ tên người nhận hàng hoặc tên Đại lý
          minLength: 2
          maxLength: 150
          example: "Nguyễn Văn An"
        customer_phone:
          type: string
          description: Số điện thoại người nhận hàng (hỗ trợ cả dạng 090... và +8490...)
          pattern: "^(0|\\+84)[35789][0-9]{8}$"
          example: "0901234567"
        customer_email:
          type: string
          format: email
          nullable: true
          description: Địa chỉ email của khách hàng
          example: "nguyenvanan@example.com"
        company_name:
          type: string
          nullable: true
          description: Tên doanh nghiệp đại lý xuất hóa đơn VAT (nếu có)
          example: "Công Ty TNHH Thú Cưng PetCare"
        vat_number:
          type: string
          nullable: true
          description: Mã số thuế doanh nghiệp xuất hóa đơn VAT (nếu có)
          example: "0312345678"
        shipping_address:
          type: string
          description: Địa chỉ giao hàng chi tiết đầy đủ
          minLength: 5
          maxLength: 500
          example: "123 Đường Nguyễn Trãi, Phường 2, Quận 5, TP. Hồ Chí Minh"
        items:
          type: array
          description: Danh sách các sản phẩm đặt mua (tối thiểu 1 sản phẩm)
          minItems: 1
          items:
            $ref: '#/components/schemas/StandardOrderItem'
        subtotal_amount:
          type: number
          minimum: 0
          nullable: true
          description: Tổng tiền niêm yết trước chiết khấu (VND)
          example: 340000
        discount_amount:
          type: number
          minimum: 0
          default: 0
          description: Tổng số tiền được giảm giá / chiết khấu (VND)
          example: 20000
        shipping_fee:
          type: number
          minimum: 0
          default: 0
          description: Phí vận chuyển giao hàng (VND, mặc định 0đ nếu miễn phí ship)
          example: 30000
        total_amount:
          type: number
          minimum: 0
          description: Tổng số tiền thực tế khách phải thanh toán cuối cùng (VND)
          example: 350000
        voucher_code:
          type: string
          nullable: true
          maxLength: 50
          description: Mã voucher khuyến mãi (nếu có)
          example: "LAPETHE2026"
        payment_method:
          type: string
          enum: [cod, bank_transfer, debt, vnpay, momo]
          default: cod
          description: Hình thức thanh toán
          example: "cod"
        note:
          type: string
          nullable: true
          maxLength: 1000
          description: Ghi chú hoặc lời nhắn giao hàng
          example: "Giao trong giờ hành chính, gọi trước khi giao 15 phút"

    # ── DÒNG SẢN PHẨM (ORDER ITEM) ──────────────────────────────────────────
    StandardOrderItem:
      type: object
      required:
        - sku
        - quantity
        - price
      properties:
        sku:
          type: string
          description: Mã SKU sản phẩm (khớp `default_code` trong Odoo)
          minLength: 1
          maxLength: 100
          example: "PAT-LAMB-100"
        product_name:
          type: string
          nullable: true
          description: Tên sản phẩm hiển thị
          example: "Pate Thịt Cừu Cho Chó 100g"
        odoo_product_id:
          type: integer
          nullable: true
          description: ID sản phẩm Odoo nếu Web Bán Hàng đã có sẵn
          example: 1042
        quantity:
          type: integer
          description: Số lượng đặt mua (nguyên dương >= 1)
          minimum: 1
          example: 10
        price:
          type: number
          description: Đơn giá sản phẩm sau chiết khấu (VND, >= 0)
          minimum: 0
          example: 25000
        original_price:
          type: number
          minimum: 0
          nullable: true
          description: Giá niêm yết gốc trước giảm
          example: 28000
        discount:
          type: number
          minimum: 0
          maximum: 100
          default: 0
          description: Tỷ lệ chiết khấu theo dòng (%)
          example: 0

    # ── PHẢN HỒI THÀNH CÔNG (RESPONSE 201) ──────────────────────────────────
    OrderCreatedResponse:
      type: object
      required:
        - success
        - message
        - data
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Đơn hàng đã được tiếp nhận thành công và đang được phân luồng xử lý"
        data:
          type: object
          required:
            - order_code
            - tracking_id
            - status
            - received_at
          properties:
            order_code:
              type: string
              example: "WEB-20260909-8A1B"
            tracking_id:
              type: string
              format: uuid
              example: "e58b14a2-97b1-4f8a-9366-4c4f3c830912"
            status:
              type: string
              enum: [pending, queued]
              example: "pending"
            received_at:
              type: string
              format: date-time
              example: "2026-09-09T09:45:00.000Z"

    # ── TIẾN ĐỘ ĐƠN HÀNG CHI TIẾT (ORDER STATUS RESPONSE) ───────────────────
    OrderStatusDetailResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            order_code:
              type: string
              example: "WEB-20260909-8A1B"
            tracking_id:
              type: string
              example: "e58b14a2-97b1-4f8a-9366-4c4f3c830912"
            status:
              type: string
              enum: [pending, queued, odoo_synced, confirmed, shipping, delivered, cancelled]
              example: "odoo_synced"
            odoo_order_id:
              type: integer
              nullable: true
              example: 2304
            odoo_order_name:
              type: string
              nullable: true
              example: "S02304"
            zalo_notified:
              type: boolean
              example: true
            total_amount:
              type: number
              example: 350000
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time

    # ── PHẢN HỒI LỖI VALIDATION (RESPONSE 400) ──────────────────────────────
    ValidationErrorResponse:
      type: object
      required:
        - success
        - error
        - details
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Dữ liệu đơn hàng không hợp lệ"
        details:
          type: array
          items:
            type: object
            required:
              - field
              - message
            properties:
              field:
                type: string
                example: "customer_phone"
              message:
                type: string
                example: "Số điện thoại không đúng định dạng số điện thoại Việt Nam"

    # ── DANH SÁCH SẢN PHẨM (PRODUCT LIST RESPONSE) ──────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Không tìm thấy dữ liệu yêu cầu"
    InternalErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Lỗi hệ thống máy chủ nội bộ"